React Kanban: Reverse engineer
Holy crap, that was a lot of work. I took a random Kanban board that
included react-beautiful-dnd and removed everything until it was a
bare-bones drag and drop app. Then I hooked it into FlowChamp and now
it's on its way to being fully functional!

TODO: Finish getting dropping hooked up