Course Blocks: Fix add/delete bugs
When adding a block, the course added to the flowchart wouldn't include
a _id because it wasn't adding the course that the API returns. It now
has the full course and deletes successfully every time.

Also, there was an issue with deleting a course where popPopup() wasn't
called in time (The course deletion would interrupt it), causing a weird
state where the popup was still in the stack when it shouldn't have
been. This no longer happens.